requirements.txt
#requests
#azure-functions
jira
azure-functions
azure-appconfiguration
